Polish apache/dubbo#4830 : [Refactor] Rename the class name of ServiceAnnotationBeanPostProcessor
